+++ lib/login_input.dart
@@ -1,0 +1,2 @@
+// TODO(victor): adicionar TextInputs com estilos iguais aos que
+// estão nos apps da Quark